:rocket: 테스트 서버 배포를 위한 CI/CD 설정
