--- backend/techpick-collector/src/main/java/techpick/TechPickCollectorApplication.java
@@ -1,16 +1,0 @@
-package techpick;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-
-@SpringBootApplication(
-	exclude = {DataSourceAutoConfiguration.class, HibernateJpaAutoConfiguration.class}
-)
-public class TechPickCollectorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TechPickCollectorApplication.class, args);
-	}
-}
