--- backend/techpick-collector/src/main/java/techpick/collector/dto/MessageDto.java
@@ -1,22 +1,0 @@
-package techpick.collector.dto;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@ToString
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MessageDto {
-
-	private String title;
-	private String message;
-
-	@Builder
-	public MessageDto(String title, String message) {
-		this.title = title;
-		this.message = message;
-	}
-}
