--- backend/techpick-collector/src/main/java/techpick/collector/repository/PickPickRepository.java
@@ -1,10 +1,0 @@
-package techpick.collector.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-import techpick.collector.entity.PickPick;
-
-@Repository
-public interface PickPickRepository extends MongoRepository<PickPick, String> {
-}
