--- backend/techpick-collector/src/main/java/techpick/collector/test/NewTask.java
@@ -1,30 +1,0 @@
-package techpick.collector.test;
-
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.Connection;
-import com.rabbitmq.client.ConnectionFactory;
-import com.rabbitmq.client.MessageProperties;
-
-public class NewTask {
-
-	private static final String TASK_QUEUE_NAME = "task_queue";
-
-	public static void main(String[] argv) throws Exception {
-		ConnectionFactory factory = new ConnectionFactory();
-		factory.setHost("localhost");
-		factory.setUsername("root");
-		factory.setPassword("1234");
-
-		try (Connection connection = factory.newConnection(); Channel channel = connection.createChannel()) {
-			channel.queueDeclare(TASK_QUEUE_NAME, true, false, false, null);
-
-			String message = String.join(" ", argv);
-
-			channel.basicPublish("", TASK_QUEUE_NAME,
-				MessageProperties.PERSISTENT_TEXT_PLAIN,
-				message.getBytes("UTF-8"));
-			System.out.println(" [x] Sent '" + message + "'");
-		}
-	}
-
-}
