--- backend/techpick-collector/src/main/java/techpick/collector/test/ReceiveLogs.java
@@ -1,31 +1,0 @@
-package techpick.collector.test;
-
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.Connection;
-import com.rabbitmq.client.ConnectionFactory;
-import com.rabbitmq.client.DeliverCallback;
-
-public class ReceiveLogs {
-	private static final String EXCHANGE_NAME = "logs";
-
-	public static void main(String[] argv) throws Exception {
-		ConnectionFactory factory = new ConnectionFactory();
-		factory.setHost("localhost");
-		factory.setUsername("root");
-		factory.setPassword("1234");
-		Connection connection = factory.newConnection();
-		Channel channel = connection.createChannel();
-
-		channel.exchangeDeclare(EXCHANGE_NAME, "fanout");
-		String queueName = channel.queueDeclare().getQueue();
-		channel.queueBind(queueName, EXCHANGE_NAME, "");
-
-		System.out.println(" [*] Waiting for messages. To exit press CTRL+C");
-
-		DeliverCallback deliverCallback = (consumerTag, delivery) -> {
-			String message = new String(delivery.getBody(), "UTF-8");
-			System.out.println(" [x] Received '" + message + "'");
-		};
-		channel.basicConsume(queueName, true, deliverCallback, consumerTag -> { });
-	}
-}
