--- backend/techpick-collector/src/main/java/techpick/collector/test/Recv.java
@@ -1,30 +1,0 @@
-package techpick.collector.test;
-
-import com.rabbitmq.client.Channel;
-import com.rabbitmq.client.Connection;
-import com.rabbitmq.client.ConnectionFactory;
-import com.rabbitmq.client.DeliverCallback;
-
-public class Recv {
-
-	private final static String QUEUE_NAME ="hello";
-
-	public static void main(String[] args) throws Exception {
-		ConnectionFactory factory = new ConnectionFactory();
-		factory.setHost("localhost");
-		factory.setUsername("root");
-		factory.setPassword("1234");
-
-		Connection connection = factory.newConnection();
-		Channel channel = connection.createChannel();
-
-		channel.queueDeclare(QUEUE_NAME, false, false, false, null);
-		System.out.println(" [*] Waiting for messages. To exit press CTRL+C");
-
-		DeliverCallback deliverCallback = (consumerTag, delivery) -> {
-			String message = new String(delivery.getBody(), "UTF-8");
-			System.out.println(" [x] Received '" + message + "'");
-		};
-		channel.basicConsume(QUEUE_NAME, true, deliverCallback, consumerTag -> {});
-	}
-}
