+++ backend/techpick-ranking/Dockerfile
@@ -1,0 +1,13 @@
+FROM amazoncorretto:17
+
+LABEL authors="TechPick"
+
+ARG JAR_FILE=build/libs/*.jar
+
+WORKDIR /app
+
+COPY ${JAR_FILE} /app/techpick-ranking.jar
+
+EXPOSE 8080
+
+ENTRYPOINT ["java", "-jar", "-Duser.timezone=Asia/Seoul", "/app/techpick-ranking.jar"]
