+++ backend/techpick-ranking/src/main/java/techpick/TechPickRankingApplication.java
@@ -1,0 +1,16 @@
+package techpick;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
+import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
+
+@SpringBootApplication(
+    exclude = {DataSourceAutoConfiguration.class, HibernateJpaAutoConfiguration.class}
+)
+public class TechPickRankingApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TechPickRankingApplication.class, args);
+    }
+}
