+++ backend/techpick-ranking/src/main/java/techpick/ranking/dto/MessageDto.java
@@ -1,0 +1,22 @@
+package techpick.ranking.dto;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@ToString
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MessageDto {
+
+    private String title;
+    private String message;
+
+    @Builder
+    public MessageDto(String title, String message) {
+        this.title = title;
+        this.message = message;
+    }
+}
