+++ backend/techpick-ranking/src/main/java/techpick/ranking/entity/Message.java
@@ -1,0 +1,26 @@
+package techpick.ranking.entity;
+
+import org.springframework.data.mongodb.core.mapping.Document;
+
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Document(collection = "message")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Message {
+
+    @Id
+    private String id;
+    private String title;
+    private String message;
+
+    @Builder
+    private Message(String title, String message) {
+        this.title = title;
+        this.message = message;
+    }
+}
