+++ backend/techpick-ranking/src/main/java/techpick/ranking/repository/PickPickRepository.java
@@ -1,0 +1,10 @@
+package techpick.ranking.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+import techpick.ranking.entity.PickPick;
+
+@Repository
+public interface PickPickRepository extends MongoRepository<PickPick, String> {
+}
