+++ backend/techpick-ranking/src/main/java/techpick/ranking/test/EmitLog.java
@@ -1,0 +1,29 @@
+package techpick.ranking.test;
+
+import java.nio.charset.StandardCharsets;
+
+import com.rabbitmq.client.Channel;
+import com.rabbitmq.client.Connection;
+import com.rabbitmq.client.ConnectionFactory;
+
+public class EmitLog {
+
+    private static final String EXCHANGE_NAME = "logs";
+
+    public static void main(String[] argv) throws Exception {
+        ConnectionFactory factory = new ConnectionFactory();
+        factory.setHost("localhost");
+        factory.setUsername("root");
+        factory.setPassword("1234");
+
+        try (Connection connection = factory.newConnection(); Channel channel = connection.createChannel()) {
+            channel.exchangeDeclare(EXCHANGE_NAME, "fanout");
+
+            String message = argv.length < 1 ? "info: Hello World!" :
+                String.join(" ", argv);
+
+            channel.basicPublish(EXCHANGE_NAME, "", null, message.getBytes(StandardCharsets.UTF_8));
+            System.out.println(" [x] Sent '" + message + "'");
+        }
+    }
+}
