+++ backend/techpick-ranking/src/main/resources/application.yaml
@@ -1,0 +1,53 @@
+# 로컬에서 techpick-api와 동시에 켜야 할 경우
+# 아래 주석을 해제하세요
+#server:
+#  port: 8082
+
+spring:
+  application:
+    name: techpick-ranking
+  output:
+    ansi:
+      enabled: always
+  sql:
+    init:
+      mode: never # schema.sql 실행시 always 키고 실행하시면 됩니다. option: never, always
+
+#  jpa:
+#    properties:
+#      hibernate:
+#        dialect: org.hibernate.dialect.MySQL8Dialect
+#    open-in-view: false
+---
+
+spring:
+  config:
+    activate:
+      on-profile: local
+  data:
+    mongodb:
+      uri: mongodb://${DOCKER_MONGO_USERNAME}:${DOCKER_MONGO_PASSWORD}@${LOCAL_MONGO_URL}/${DOCKER_MONGO_DATABASE}?authSource=${DOCKER_MONGO_AUTH}
+      database: ${DOCKER_MONGO_DATABASE}
+      username: ${DOCKER_MONGO_USERNAME}
+      password: ${DOCKER_MONGO_PASSWORD}
+
+  rabbitmq:
+    host: ${RABBITMQ_HOST}
+    port: ${RABBITMQ_PORT}
+    username: ${RABBITMQ_USERNAME}
+    password: ${RABBITMQ_PASSWORD}
+
+
+#  jpa:
+#    properties:
+#      hibernate:
+#        format_sql: false
+#        show_sql: true
+#    hibernate:
+#      ddl-auto: create
+#  datasource:
+#    url:
+#    username:
+#    password:
+#    driver-class-name: com.mysql.cj.jdbc.Driver
+
